Add "y" attributes to <ITEM> tags
The year information comes from the main table (`main.csv`), which is
not included in this repository. Thus, this information needs to be
included in the XML as it's not available anywhere else.

On the other hand, the place and collector information are provided
in the CSV files (`laul_koguja.csv`, `laul_koht.csv`), so no need to
duplicate that in the XML.

This commit changes both the code for producing XML and the resulting
XML files.